fix MovieDetailBase

+++ src/schemas/movies.py
@@ -63,7 +63,7 @@ class MovieDetailBase(BaseModel):
     status: str
     orig_lang: str
     budget: int
-    revenue: int
+    revenue: float
     country: str
